[worker/tests] added function that returns a sample encrypted msg for testing
